Progress to making a complete game
Added a few new obstacles. Objects are now color-coded, with dark objects blocking projectiles and light objects not blocking them. Expanded the total number of levels to 23, with several tutorial levels. The level select screen now includes spaces to select 36 different levels. Added music and sound effects, controlled by a script. When the player walked and then immediately punched, the punch animation was delayed, and this has now been resolved. Several more levels and sound effects are due to be added very soon, as well as a full credits page.