Android Update in-progress 1
Made many many changes! Now using a new font and button style. Non-button text now includes a grey border to set it apart from the background. Zoomed in the camera to better view on mobile, and had to redesign all of the levels to accomodate that change. Added touch controls, specifically a joystick and punch/roll buttons. Moved many UI elements around on all canvases. Added a settings canvas which currently has buttons to mute the game and swap the controls between left-handed and right-handed modes. Added a thicker transparent barrier on the outside of the walls so the player is less likely to pass through them.