Android Update-in-progress 3
Added several different colored materials for the player, to be unlocked by accomplishing various achievements. Improved the Credits readability. Added the Achievement page, accessible from the main menu, which shows a list of the available achievements, what colors they unlock, the ability to toggle those colors, and also the color as applied to a little mini version of the player model. Improved the Settings screen, which is now layed out much cleaner and features the ability to toggle if the game controls appear on the main menu.  Added the ability to save the game in between play sessions. Added the icons for the punch and roll buttons.