Android Update-in-progress 5
Decided on the name "Sock 'n Roll". Created the game's app icons. Created banner art to be displayed on the home page, showing the name. Removed all references to "Battle Box", as well as removing references to the Desktop controls. Sock 'n Roll is now destined to be exclusively a mobile game. Cleaned up some code and unused assets.