unpublished change - post release 1.01
Added UI scaling to make all of the canvases vary in size with aspect ratio, with the result of making all UI elements fit cleanly on the screen. Also made different UI elements use a position relative to the top or bottom of the screen, which keeps the screen layout constant across devices. Sock 'n Roll is now locked to being played in landscape mode, with possible screen rotation.